Make example prompt consistent with Quick-start context (#2215)
The goal of this doc is to provide an example where the triage agent can
hand off the question to either a math tutor or a history tutor.
However, the example prompt here has nothing to do with history or math;
as a result, agent handoff will be skipped, which might confuse Agent
SDK beginners.

Also, in the "Put it all together" section, the prompt used is different
from the example prompt here.

This commit is updating the example prompt to be a history-related
question, making it consistent with the section below.